requirements.txt
django
django-crispy-forms
Pillow
gunicorn
psycopg2
django-phone-field
phonenumbers